app/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
tensorflow==2.13.0
numpy==1.24.3
pandas
pillow
pyarrow
django-cors-headers
django-filter
djangorestframework
markdown